requirements.txt
beautifulsoup4
psycopg2-binary
selenium
Pillow
djangorestframework
pandas
numpy
django-pandas
seaborn
matplotlib
ipykernel